toggle-box: Support Block API Version 2
