+++ blocks/src/block/toggle-box/index.js
@@ -1,0 +1,28 @@
+/**
+ * Cocoon Blocks
+ * @author: yhira
+ * @link: https://wp-cocoon.com/
+ * @license: http://www.gnu.org/licenses/gpl-2.0.html GPL v2 or later
+ */
+
+import {THEME_NAME} from '../../helpers';
+import { __ } from '@wordpress/i18n';
+
+import edit from './edit';
+import save from './save';
+import deprecated from './deprecated';
+import metadata from './block.json';
+
+const { name } = metadata;
+export { metadata, name };
+
+export const settings = {
+  title: __( 'トグルボックス', THEME_NAME ),
+  icon: 'randomize',
+  description: __( 'クリックすることでコンテンツ内容の表示を切り替えることができるボックスです。', THEME_NAME ),
+  example: {},
+
+  edit,
+  save,
+  deprecated,
+};
