+++ blocks/src/blocks.js
@@ -23,6 +23,7 @@ import * as infoBox from './block/info-box';
 import * as searchBox from './block/search-box';
 import * as stickyBox from './block/sticky-box';
 import * as tabBox from './block/tab-box';
+import * as toggleBox from './block/toggle-box';
 
 const cocoonBlocks = [
   balloon,
@@ -36,6 +37,7 @@ const cocoonBlocks = [
   searchBox,
   stickyBox,
   tabBox,
+  toggleBox,
 ];
 
 
@@ -88,7 +90,7 @@ registerCocoonBlocks();
 //import './block/blogcard/block.js';
 //import './block/button/block.js';
 //import './block/button-wrap/block.js';
-import './block/toggle-box/block.js';
+//import './block/toggle-box/block.js';
 //import './block/search-box/block.js';
 import './block/timeline/block.js';
 //import './block/icon-list/block.js';
